fixed small Bugs in function to extract sequences from the dbmart array

+++ utils/sequencing.h
@@ -18,9 +18,9 @@ struct temporalSequence{
     unsigned int patientID;
 }__attribute__((packed));
 
-int extractSequencesFromArray(dbMartEntry * dbMart, size_t numOfPatients, const size_t * startPositions,
+size_t extractSequencesFromArray(dbMartEntry * dbMart, size_t numOfPatients, const size_t * startPositions,
                               size_t numberOfDbMartEntries,  const std::string& outPutDirectory,
-                              const std::string& outputFilePrefix, int patIDLength = 7, int numOfThreads =1);
+                              const std::string& outputFilePrefix, int patIDLength = 7, int numOfThreads = 1);
 
 unsigned int getBucket(unsigned int min, unsigned int max, int threshold, unsigned int duration);
